# public/openapi.yaml
openapi: 3.0.3
info:
  title: ApplyCopilot REST API (Phase 1)
  description: API Contracts for Authentication and CV Import pipelines in ApplyCopilot V2.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Development Server
paths:
...
  /api/profile/summaries/generate:
    post:
      summary: Generate AI profile summary
      description: Generates a professional summary using LLM based on user profile context, limited to 10 requests per user per hour.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tone
              properties:
                tone:
                  type: string
                  example: professional
                instructions:
                  type: string
                  example: Focus on backend scaling experience.
      responses:
        '200':
          description: Summary generated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                    example: Staff Software Engineer
                  content:
                    type: string
                    example: Experienced software leader specializing in high-performance cloud architectures.
        '400':
          description: Invalid input or missing tone.
        '401':
          description: Unauthorized.
        '429':
          description: Rate limit reached (max 10 requests per hour).
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: Session cookie generated by NextAuth.js